requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
requests==2.31.0
faiss-cpu==1.7.4
sentence-transformers==2.2.2
python-dotenv==1.0.0
numpy==1.24.3
beautifulsoup4==4.12.2
aiohttp==3.9.1
lxml==4.9.3
selenium==4.16.0
webdriver-manager==4.0.1 